+++ src/components/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app/app';
-import {Setting} from './constants';
+import {Setting} from '../constants';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
